refactor common helpers
